Etc. Slide Module 클래스 분리 및 리뷰사항 반영
이벤트리스너 진입을 금지하는 flag변수 이름 stopFlag로 변경
Slide 클래스와 ImageCarousel 클래스 파일로 분리
DOM에 저장하던 data-state를 자바스크립트에서 관리하는 state로 변경(Slide
인스턴스가 관리)
