+++ README.md
@@ -1,1 +1,52 @@
 # fe-kakaopage
+
+```Directory
+├── 📂assets
+│   └── ...
+├── 📂json
+│   └── ...
+├── 📂css
+│   ├── 📂pageHeader
+│   │   ├── 📘header.css
+│   │   └── 📘category.css
+│   ├── 📂pageMain
+│   │   ├── 📘banner.css
+│   │   ├── 📘category.css
+│   │   ├── 📘main.css
+│   │   ├── 📘preview.css
+│   │   ├── 📘taglist.css
+│   │   ├── 📘webtoonComponent.css
+│   │   └── 📘webtoonFilter.css
+│   ├── 📘reset.css
+│   └── 📘var.css
+├── 📂js
+│   ├── 📂dayFilter
+│   │   ├── 📒addClickEvent.js
+│   │   ├── 📒clearDayFilter.js
+│   │   ├── 📒drawDayFilter.js
+│   │   ├── 📒initMainCategoryDay.js
+│   │   ├── 📒replaceDayContents.js
+│   │   └── 📒setDayFiilter.js
+│   ├── 📂preview
+│   │   ├── 📒clearPreviews.js
+│   │   ├── 📒clearSlideBtn.js
+│   │   ├── 📒createPreviewEl.js
+│   │   ├── 📒drawPreviews.js
+│   │   ├── 📒drawSlideBtn.js
+│   │   └── 📒setPreview.js
+│   ├── 📂tagList
+│   │   ├── 📒clearTagListEl.js
+│   │   ├── 📒drawTagListEl.js
+│   │   └── 📒setTagListEl.js
+│   ├── 📂webtoonComponent
+│   │   ├── 📒clearWebtoonContents.js
+│   │   ├── 📒drawWebtoonContents.js
+│   │   ├── 📒setWebtoonContents.js
+│   │   ├── 📒webtoonComponent.js
+│   │   └── 📒cover.js
+│   ├── 📒pageHeaderCategory.js
+│   ├── 📒pageHeaderSearchBar.js
+│   └── 📒pageMainCategory.js
+├── 📄index.html
+└── 📘styles.css
+```
