+++ SlideModule/index.html
@@ -18,6 +18,6 @@
       <button class="prev-btn">&lt;</button>
       <button class="next-btn">&gt;</button>
     </div>
-    <script src="./app.js"></script>
+    <script src="./app.js" type="module"></script>
   </body>
 </html>
